--- components/Icons/Icon.tsx
@@ -1,10 +1,0 @@
-import { styled } from "../../lib/styles/stitches.config";
-
-const Icon = styled("svg", {
-  width: "1.25em",
-  height: "1.25em",
-  verticalAlign: "-0.3em",
-  fill: "currentColor",
-});
-
-export default Icon;
